+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated content report style based on the latest NIPs and fixed some bugs with reporting.
 - Fixed the launch screen layout on iPad
+- Multiple consecutive newlines will be replaced by a single new line in note content.
+- Leading and trailing whitespace will no longer be rendered in note content.
+- Removed the screen to fill out profile information from onboarding and replaced it with a call to action in the sidebar. 
+
+## [0.1 (59)] - 2023-07-21Z
+
+- Add a loading placeholder for note contents.
+- Fixed several crashes.
 - Added Dutch, Japanese, and Persian translations. Thanks matata, yutaro, and eru-desu! 
 - Added some visual artists to the list of featured users.